Fix invisible animals in cages in multiplayer
BroadcastPacket takes a list of players to EXCLUDE
SendPacket takes a list of players to INCLUDE
Excluing the joined player meant the player never received the server config which resulted in cages being invisible

Additionally no need to re-register the channel. Just do it once in the universal Start function.